pyproject.toml
[project]
name = "minimaker"
version = "0.1.0"
description = "Minimal LLM training framework for single-node multi-GPU"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.2.0",
    "hydra-core>=1.3",
    "omegaconf>=2.3",
    "wandb>=0.16",
    "tensorboard>=2.15",
    "datasets>=2.16",
    "transformers>=4.37",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "ruff>=0.2"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/minimaker"]

[tool.ruff]
line-length = 100
target-version = "py310"
